+++ js/app.js
@@ -302,6 +302,8 @@ loadCubeTexture('teatro', '/assets/textures/cube/teatro/', '.png') // 2
 // global.textures[n]
 loadTexture('StoneDiff', '/assets/textures/stone_tiles_02_diff_1k', '.jpg') // 0
 loadTexture('StoneDisp', '/assets/textures/stone_tiles_02_disp_4k', '.png') // 1
+loadTexture('Facce_colori', '/assets/textures/Mosaico_facce_colori', '.jpg') // 2
+loadTexture('Facce_bk', '/assets/textures/Mosaico_facce_bk', '.jpg') // 3
 
 
 // INIT
