--- sketch/3/5.js
@@ -1,123 +1,0 @@
-// Marching cubes // prova piccoli blob
-
-import Stats from 'three/addons/libs/stats.module.js' // XXX
-import { OrbitControls } from 'three/addons/controls/OrbitControls.js'
-
-import { MarchingCubes } from 'three/addons/objects/MarchingCubes.js'
-
-let scene
-let effect
-let material
-let animation
-let onWindowResize
-let controls
-
-export function sketch() {
-    console.log("Sketch launched")
-    const stats = new Stats() // XXX
-    canvas3D.appendChild(stats.dom)
-
-    let time = 0
-    const clock = new THREE.Clock()
-
-    // CAMERA
-    let camera = new THREE.PerspectiveCamera(50, window.innerWidth / window.innerHeight, 0.1, 1000)
-    camera.position.z = 60 // allontanato la camera
-
-    // WINDOW RESIZE
-    const onWindowResize = () => {
-        camera.aspect = window.innerWidth / window.innerHeight
-        camera.updateProjectionMatrix()
-        renderer.setSize(window.innerWidth, window.innerHeight)
-    }
-    window.addEventListener('resize', onWindowResize)
-
-    // CONTROLS
-    controls = new OrbitControls(camera, renderer.domElement)
-
-    // SCENE
-    scene = new THREE.Scene()
-    material = new THREE.MeshStandardMaterial({ color: 0xaaaaff, envMap: global.cubeTextures[1].texture, roughness: 0, metalness: 1 })
-    let resolution = 28;
-    // effect
-    let effectController = {
-        speed: .8,
-        numBlobs: 40,
-        resolution: 70,
-        isolation: 80,
-        floor: false,
-        wallx: false,
-        wallz: false,
-        dummy: function () { }
-    }
-    effect = new MarchingCubes(resolution, material, true, true, 100000)
-    effect.position.set(0, 0, 0)
-    effect.scale.set(50, 50, 50)
-    effect.enableUvs = false
-    effect.enableColors = false
-    scene.add(effect)
-    // this controls content of marching cubes voxel field
-    function updateCubes(object, time, numblobs, floor, wallx, wallz) {
-        object.reset()
-        // fill the field with some metaballs
-        const subtract = 12;
-        const strength = .2 / ((Math.sqrt(numblobs) - 1) / 4 + 1)
-        for (let i = 0; i < numblobs; i++) {
-            const ballx = Math.sin(i + 1.26 * time * (1.03 + 0.5 * Math.cos(0.21 * i))) * 0.27 + 0.5
-            const bally = Math.abs(Math.cos(i + 1.12 * time * Math.cos(1.22 + 0.1424 * i))) * 0.77 // dip into the floor
-            const ballz = Math.cos(i + 1.32 * time * 0.1 * Math.sin((0.92 + 0.53 * i))) * 0.27 + 0.5
-            object.addBall(ballx, bally, ballz, strength, subtract)
-        }
-        if (floor) object.addPlaneY(2, 12)
-        if (wallz) object.addPlaneZ(2, 12)
-        if (wallx) object.addPlaneX(2, 12)
-        object.update()
-    }
-
-    // NOISE
-    // noise3D = NOISE.createNoise3D()
-    // const t0 = Math.random() * 10
-
-    // LIGHTS
-    const light = new THREE.DirectionalLight(0xffffff, 1)
-    light.position.set(0.5, 0.5, 1)
-    scene.add(light)
-    const pointLight = new THREE.PointLight(0x0000ff, 10, 0)
-    pointLight.position.set(0, 0, 100)
-    scene.add(pointLight)
-    const ambientLight = new THREE.AmbientLight(0x00faff, 100)
-    scene.add(ambientLight)
-
-
-    // ANIMATE
-    const animate = () => {
-        stats.begin() // XXX
-
-        // ANIMATION
-        const delta = clock.getDelta();
-        time += delta * effectController.speed * 0.5; 
-        // marching cubes
-        if (effectController.resolution !== resolution) {
-            resolution = effectController.resolution;
-            effect.init(Math.floor(resolution));
-        }
-        if (effectController.isolation !== effect.isolation) {
-            effect.isolation = effectController.isolation;
-        }
-        updateCubes(effect, time, effectController.numBlobs, effectController.floor, effectController.wallx, effectController.wallz);
-        // ...
-
-        renderer.render(scene, camera) // RENDER
-        stats.end() // XXX
-
-        animation = requestAnimationFrame(animate) // CIAK
-    }
-    animate()
-}
-
-export function dispose() {
-    cancelAnimationFrame(animation)
-    controls?.dispose()
-    material?.dispose()
-    window.removeEventListener('resize', onWindowResize)
-}
